Add a lot of new features for Pavlov game
This commit adds new functionality to Pavlov lobby
browser to browse lobbies and players data including
listing players in a lobby etc.

This uses hidden details of EOS SDK that's not exposed
in the official SDK but is sent over the wire and few
extra API calls to fetch user data.

Players are presented in embedded WebView control
with their avatars and data.