Fixed a bug with restoring WebView & added navigation-to-profile
This commit fixes a bug where a WebView for lobby details got corrupted
(stopped being displayed) after minimizing window to the tray icon and
restoring it. The fix is to recreate webview in its place after the
main window was restored. Root cause is unknown but it's either a
wxWidgets or webview2 bug/quirk.

Apart from that, a new small feature was added where by clicking on
platform icon (e.g. Steam icon) on details panel, that player's profile
page will be opened in default system browser. For now this is
implemented only for Steam/PCVR players in Pavlov game.